requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11.0.3

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Image Processing
pillow==10.1.0
opencv-python==4.8.1.78
mediapipe==0.10.21

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.14.0

# Additional dependencies
numpy==1.24.4
